+++ media/scripts/main.js
@@ -7,7 +7,7 @@ import { updateViews } from './ui/viewManager.js';
 
 (function () {
   document.addEventListener('DOMContentLoaded', () => {
-    // console.log('main.js loaded');
+    console.log('main.js loaded');
 
     window.currentState = {
       view: 'saved-colors',
@@ -15,36 +15,9 @@ import { updateViews } from './ui/viewManager.js';
       projects: [],
       currentProject: null,
     };
-    window.serverReady = false;
     window.vscode = acquireVsCodeApi();
 
     vscode.postMessage({ command: 'ready' });
-    function startServerPing() {
-      let attempts = 0;
-      const maxAttempts = 10; 
-
-      const intervalId = setInterval(async () => {
-        attempts++;
-        try {
-          const res = await fetch('http://localhost:3000/api/ping');
-          if (res.ok) {
-            window.serverReady = true;
-            clearInterval(intervalId);
-            document.dispatchEvent(new Event('serverReady'));
-          }
-        } catch (err) {
-          // Do nothing if ping fails
-        }
-
-        if (attempts >= maxAttempts) {
-          clearInterval(intervalId);
-        }
-      }, 3000);
-    }
-   startServerPing();
-
-    
-
     initNavigation();
     initSavedColors();
     initProjects();
@@ -60,8 +33,6 @@ import { updateViews } from './ui/viewManager.js';
       }
     });
     document.addEventListener('stateUpdated', updateViews);
-    document.addEventListener('serverReady', () => {
-      updateViews();
-    });
+
   });
 })();