+++ src/utils/types.ts
@@ -3,15 +3,3 @@ export interface ColorProject {
   name: string;
   colors: string[];
 }
-export interface ThemeCollection {
-  mainColor: string;
-  themes: Theme[];
-}
-export interface Theme {
-  text: string;
-  name: string;
-  background: string;
-  accent: string;
-  description: string;
-  role: string;
-}