+++ CHANGELOG.md
@@ -1,9 +1,22 @@
-# Change Log
+# Changelog
 
-All notable changes to the "color-store" extension will be documented in this file.
+All notable changes to the **Color Store** extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),  
+and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.0.2] - 2025-06-11
 
-- Initial release
+### Added
+
+- Preview multiple shades of a selected color
+
+## [0.0.1] - 2025-06-10
+
+### Added
+
+- Initial public release of **Color Store**
+- Generate and preview color theme suggestions based on a selected color
+- Save and organize favorite colors
+- Copy colors in multiple formats (Tailwind, CSS, plain hex, etc.)
+- Project-based color organization UI